requirements.txt
alpaca-py >= 0.13.4
matplotlib >= 3.8.2